requirements.txt
flask==3.1.1
flask-restx==1.3.0
flask_cors==6.0.1
pymongo==4.13.2
werkzeug==3.1.3
python-dotenv==1.1.1
flask-jwt-extended==4.7.1
email-validator==2.2.0
bcrypt==4.3.0
Flask-APScheduler==1.13.1
boto3==1.39.8
flask_cors==6.0.1
python-dateutil==2.9.0
